Rename Jbot.py to Jbot.py (OUTDATED)
Updated, this code works, however the bot is currently running on much more updated code (and getting more commands daily).  Therefore, I'm not sure how often I'll update the github code to reflect the changes.  If people want to see the latest code, I will provide it, but until then this github will mostly be a feature list with occasional version updates to the code.